+++ src/features/usage-consolidation-accordion/summary-table/column-properties.js
@@ -1,5 +1,10 @@
 import React from 'react';
-import { FormattedNumber } from 'react-intl';
+import {
+  FormattedMessage,
+  FormattedNumber,
+} from 'react-intl';
+
+import { InfoPopover } from '@folio/stripes/components';
 
 import { combineMCLProps } from '../../../components/utilities';
 import {
@@ -22,7 +27,32 @@ const DEFAULT_SUMMARY_TABLE_COLUMN_WIDTH = {
 };
 
 const DEFAULT_SUMMARY_TABLE_COLUMN_MAPPING = {
-  [DEFAULT_SUMMARY_TABLE_COLUMNS.USAGE]: 'ui-eholdings.usageConsolidation.summary.totalUsage',
+  [DEFAULT_SUMMARY_TABLE_COLUMNS.USAGE]: ({ metricType, entityType }) => ((
+    <>
+      <FormattedMessage id="ui-eholdings.usageConsolidation.summary.totalUsage" />
+      { /* eslint-disable-next-line jsx-a11y/interactive-supports-focus, jsx-a11y/click-events-have-key-events */ }
+      <span
+        role="button"
+        onClick={(e) => {
+          // We don't need to open / close the accordion by clicking on the info icon
+          e.stopPropagation();
+        }}
+      >
+        <InfoPopover
+          allowAnchorClick
+          hideOnButtonClick
+          iconSize="medium"
+          content={(
+            <FormattedMessage
+              id={`ui-eholdings.usageConsolidation.summary.totalUsage.infoPopover.${entityType}`}
+              values={{ metricType }}
+            />
+          )}
+          buttonLabel={<FormattedMessage id="ui-eholdings.usageConsolidation.infoPopover.buttonLabel" />}
+        />
+      </span>
+    </>
+  )),
   [DEFAULT_SUMMARY_TABLE_COLUMNS.COST_PER_USE]: 'ui-eholdings.usageConsolidation.summary.costPerUse',
   [DEFAULT_SUMMARY_TABLE_COLUMNS.UC_ACTIONS]: null,
 };
@@ -35,20 +65,29 @@ const DEFAULT_FORMATTER = (currency) => ({
 
 export const getCostPerUseFormatter = DEFAULT_FORMATTER;
 
-export const getSummaryTableColumnProperties = (intl, customProps = {}, currency) => {
+export const getSummaryTableColumnProperties = (intl, customProps = {}, args) => {
+
   const defaultProps = {
     visibleColumns: [...Object.values(DEFAULT_SUMMARY_TABLE_COLUMNS)],
     columnMapping: { ...DEFAULT_SUMMARY_TABLE_COLUMN_MAPPING },
     columnWidths: { ...DEFAULT_SUMMARY_TABLE_COLUMN_WIDTH },
-    formatter: DEFAULT_FORMATTER(currency),
+    formatter: DEFAULT_FORMATTER(args.currency),
   };
 
   const combinedProps = combineMCLProps(defaultProps)(customProps);
 
   const formattedColumnMappingMessages = Object.keys(combinedProps.columnMapping).reduce((memo, currentKey) => {
-    memo[currentKey] = combinedProps.columnMapping[currentKey]
-      ? intl.formatMessage({ id: combinedProps.columnMapping[currentKey] })
-      : null;
+    const columnMappingValue = combinedProps.columnMapping[currentKey];
+
+    if (!columnMappingValue) {
+      memo[currentKey] = null;
+    } else if (typeof columnMappingValue === 'string') {
+      memo[currentKey] = intl.formatMessage({ id: columnMappingValue });
+    } else if (typeof columnMappingValue === 'function') {
+      memo[currentKey] = columnMappingValue(args);
+    } else {
+      memo[currentKey] = columnMappingValue;
+    }
 
     return memo;
   }, {});