+++ src/features/usage-consolidation-accordion/usage-consolidation-accordion.js
@@ -150,7 +150,10 @@ const UsageConsolidationAccordion = ({
       platformType,
       year,
     } = filterData;
-    const { startMonth } = usageConsolidation.data;
+    const {
+      startMonth,
+      metricType,
+    } = usageConsolidation.data;
 
     if (!isCostPerUseDataLoaded && !isCostPerUseDataLoadingFailed) {
       return null;
@@ -171,6 +174,7 @@ const UsageConsolidationAccordion = ({
           year={year}
           onViewTitles={handleViewTitles}
           onLoadMoreTitles={handleLoadMoreTitles}
+          metricType={metricType}
         />
       );
     } else if (recordType === entityTypes.TITLE) {
@@ -181,6 +185,7 @@ const UsageConsolidationAccordion = ({
           publicationType={publicationType}
           platformType={platformType}
           startMonth={startMonth}
+          metricType={metricType}
         />
       );
     } else if (recordType === entityTypes.RESOURCE) {
@@ -190,6 +195,7 @@ const UsageConsolidationAccordion = ({
           year={year}
           platformType={platformType}
           startMonth={startMonth}
+          metricType={metricType}
         />
       );
     }